+++ crates/resvg/src/filter/composite.rs
@@ -2,7 +2,9 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use crate::{f64_bound, FuzzyZero, ImageRef, ImageRefMut, RGBA8};
+use super::{f32_bound, ImageRef, ImageRefMut};
+use rgb::RGBA8;
+use usvg::ApproxZeroUlps;
 
 /// Performs an arithmetic composition.
 ///
@@ -12,11 +14,11 @@ use crate::{f64_bound, FuzzyZero, ImageRef, ImageRefMut, RGBA8};
 /// # Panics
 ///
 /// When `src1`, `src2` and `dest` have different sizes.
-pub fn arithmetic_composite(
-    k1: f64,
-    k2: f64,
-    k3: f64,
-    k4: f64,
+pub fn arithmetic(
+    k1: f32,
+    k2: f32,
+    k3: f32,
+    k4: f32,
     src1: ImageRef,
     src2: ImageRef,
     dest: ImageRefMut,
@@ -25,16 +27,16 @@ pub fn arithmetic_composite(
     assert!(src1.height == src2.height && src1.height == dest.height);
 
     let calc = |i1, i2, max| {
-        let i1 = i1 as f64 / 255.0;
-        let i2 = i2 as f64 / 255.0;
+        let i1 = i1 as f32 / 255.0;
+        let i2 = i2 as f32 / 255.0;
         let result = k1 * i1 * i2 + k2 * i1 + k3 * i2 + k4;
-        f64_bound(0.0, result, max)
+        f32_bound(0.0, result, max)
     };
 
     let mut i = 0;
     for (c1, c2) in src1.data.iter().zip(src2.data.iter()) {
         let a = calc(c1.a, c2.a, 1.0);
-        if a.is_fuzzy_zero() {
+        if a.approx_zero_ulps(4) {
             i += 1;
             continue;
         }