+++ crates/resvg/src/filter/morphology.rs
@@ -2,36 +2,21 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use alloc::vec;
-
-use crate::{ImageRefMut, RGBA8};
-
-/// A morphology operation.
-#[allow(missing_docs)]
-#[derive(Clone, Copy, PartialEq, Debug)]
-pub enum MorphologyOperator {
-    Erode,
-    Dilate,
-}
+use super::ImageRefMut;
+use rgb::RGBA8;
+use usvg::filter::MorphologyOperator;
 
 /// Applies a morphology filter.
 ///
 /// `src` pixels should have a **premultiplied alpha**.
 ///
-/// # Panics
-///
-/// When `rx` or `ry` is negative.
-///
 /// # Allocations
 ///
 /// This method will allocate a copy of the `src` image as a back buffer.
-pub fn morphology(operator: MorphologyOperator, rx: f64, ry: f64, src: ImageRefMut) {
-    assert!(!rx.is_sign_negative());
-    assert!(!ry.is_sign_negative());
-
+pub fn apply(operator: MorphologyOperator, rx: f32, ry: f32, src: ImageRefMut) {
     // No point in making matrix larger than image.
-    let columns = core::cmp::min(rx.ceil() as u32 * 2, src.width);
-    let rows = core::cmp::min(ry.ceil() as u32 * 2, src.height);
+    let columns = std::cmp::min(rx.ceil() as u32 * 2, src.width);
+    let rows = std::cmp::min(ry.ceil() as u32 * 2, src.height);
     let target_x = (columns as f32 / 2.0).floor() as u32;
     let target_y = (rows as f32 / 2.0).floor() as u32;
 
@@ -39,7 +24,7 @@ pub fn morphology(operator: MorphologyOperator, rx: f64, ry: f64, src: ImageRefM
     let height_max = src.height as i32 - 1;
 
     let mut buf = vec![RGBA8::default(); src.data.len()];
-    let mut buf = ImageRefMut::new(&mut buf, src.width, src.height);
+    let mut buf = ImageRefMut::new(src.width, src.height, &mut buf);
     let mut x = 0;
     let mut y = 0;
     for _ in src.data.iter() {
@@ -62,15 +47,15 @@ pub fn morphology(operator: MorphologyOperator, rx: f64, ry: f64, src: ImageRefM
 
                 let p = src.pixel_at(tx as u32, ty as u32);
                 if operator == MorphologyOperator::Erode {
-                    new_p.r = core::cmp::min(p.r, new_p.r);
-                    new_p.g = core::cmp::min(p.g, new_p.g);
-                    new_p.b = core::cmp::min(p.b, new_p.b);
-                    new_p.a = core::cmp::min(p.a, new_p.a);
+                    new_p.r = std::cmp::min(p.r, new_p.r);
+                    new_p.g = std::cmp::min(p.g, new_p.g);
+                    new_p.b = std::cmp::min(p.b, new_p.b);
+                    new_p.a = std::cmp::min(p.a, new_p.a);
                 } else {
-                    new_p.r = core::cmp::max(p.r, new_p.r);
-                    new_p.g = core::cmp::max(p.g, new_p.g);
-                    new_p.b = core::cmp::max(p.b, new_p.b);
-                    new_p.a = core::cmp::max(p.a, new_p.a);
+                    new_p.r = std::cmp::max(p.r, new_p.r);
+                    new_p.g = std::cmp::max(p.g, new_p.g);
+                    new_p.b = std::cmp::max(p.b, new_p.b);
+                    new_p.a = std::cmp::max(p.a, new_p.a);
                 }
             }
         }