+++ crates/resvg/tests/integration/main.rs
@@ -1,5 +1,5 @@
 use once_cell::sync::Lazy;
-use rgb::FromSlice;
+use rgb::{FromSlice, RGBA8};
 use usvg::{fontdb, TreeParsing, TreeTextToPath};
 
 #[rustfmt::skip]
@@ -52,7 +52,7 @@ pub fn render(name: &str) -> usize {
     // pixmap.save_png(&format!("tests/{}.png", name)).unwrap();
 
     let mut rgba = pixmap.take();
-    svgfilters::demultiply_alpha(rgba.as_mut_slice().as_rgba_mut());
+    demultiply_alpha(rgba.as_mut_slice().as_rgba_mut());
 
     let expected_data = load_png(&png_path);
     assert_eq!(expected_data.len(), rgba.len());
@@ -98,7 +98,7 @@ pub fn render_extra_with_scale(name: &str, scale: f32) -> usize {
     // pixmap.save_png(&format!("tests/{}.png", name)).unwrap();
 
     let mut rgba = pixmap.take();
-    svgfilters::demultiply_alpha(rgba.as_mut_slice().as_rgba_mut());
+    demultiply_alpha(rgba.as_mut_slice().as_rgba_mut());
 
     let expected_data = load_png(&png_path);
     assert_eq!(expected_data.len(), rgba.len());
@@ -204,4 +204,14 @@ fn gen_diff(name: &str, img1: &[u8], img2: &[u8]) -> Result<(), png::EncodingErr
     encoder.set_depth(png::BitDepth::Eight);
     let mut writer = encoder.write_header()?;
     writer.write_image_data(&img3)
+}
+
+/// Demultiplies provided pixels alpha.
+fn demultiply_alpha(data: &mut [RGBA8]) {
+    for p in data {
+        let a = p.a as f64 / 255.0;
+        p.b = (p.b as f64 / a + 0.5) as u8;
+        p.g = (p.g as f64 / a + 0.5) as u8;
+        p.r = (p.r as f64 / a + 0.5) as u8;
+    }
 }